Setup locations for media, static files and templates.
We create a resonable hierarchy for both static and templates.

For static files, put them into separate directories by type of object.

This means things can easily be extended to include pdfs, files, etc.

For templates categorise them where they are likely to be used.
This may have issues with 'pdf' and 'csv' templates which can end up
being used in both email and web views. But those are not as common.

We have 'templates/www' prioritise first so that Django will look for
403, 404 and 500 templates there first. Otherwise they will need to be
in the root ('templates') folder.

Finally, we use 'docroot' as the place to collect all static files into.

'docroot' is useful in the local and self-deployment cases. When
deploying to PaaS providers it is something you will want to
re-consider.